+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Changes are grouped as follows
 
 * In the `unstable` package: Task description
 * In the `unstable` package: Convenience methods to create scheduled tasks (ie, `from_interval`, `from_cron`)
+* In the `unstable` package: A `TaskContext` class, which exposes a logging interface to use for logging and error
+  reporting within a task.
 
 ### Changed
 
@@ -26,6 +28,7 @@ Changes are grouped as follows
   to several more specific methods, with the goal of being more convenient to use. E.g. `self.fatal(desc)` or
   `self.warning(desc)`. As well as separating out `begin_error`, `begin_warning` etc for long-lasting errors, to avoid
   the `self.warning(...).instant()` workaround.
+* In the `unstable` package: Change task signature to take a `TaskContext` as an arg
 
 ### Fixed